requirements.txt
watchdog>=4.0
PyPDF2>=3.0
google-generativeai>=0.8
streamlit>=1.38
python-dotenv>=1.0
